+++ R/var_label.R
@@ -7,7 +7,8 @@
 #' @param unlist for data frames, return a named vector instead of a list
 #' @param null_action for data frames, by default `NULL` will be returned for
 #' columns with no variable label. Use `"fill"` to populate with the column name
-#' instead, or `"skip"` to remove such values from the returned list.
+#' instead, `"skip"` to remove such values from the returned list, `"na"` to
+#' populate with `NA` or `"empty"` to populate with an empty string (`""`).
 #' @param recurse if `TRUE`, will apply `var_label()` on packed columns
 #' (see [tidyr::pack()]) to return the variable labels of each sub-column;
 #' otherwise, the label of the group of columns will be returned.
@@ -70,7 +71,8 @@ var_label.default <- function(x, ...) {
 #' @export
 var_label.data.frame <- function(x,
                                  unlist = FALSE,
-                                 null_action = c("keep", "fill", "skip"),
+                                 null_action =
+                                   c("keep", "fill", "skip", "na", "empty"),
                                  recurse = FALSE,
                                  ...) {
   if (recurse) {
@@ -95,6 +97,24 @@ var_label.data.frame <- function(x,
       r,
       names(r),
       SIMPLIFY = FALSE
+    )
+  }
+
+  if (null_action == "empty") {
+    r <- lapply(
+      r,
+      function(x) {
+        if (is.null(x)) "" else x
+      }
+    )
+  }
+
+  if (null_action == "na") {
+    r <- lapply(
+      r,
+      function(x) {
+        if (is.null(x)) as.character(NA) else x
+      }
     )
   }